use metadata username for post details overlay; fix overlay for profile posts
