+++ Cargo.toml
@@ -206,22 +206,22 @@ base64 = "0.22"
 
 yral-pump-n-dump-common = { git = "https://github.com/dolr-ai/yral-common.git", features = [
     "client",
-], branch = "master", version = "0.1.0" }
-sns-validation = { git = "https://github.com/dolr-ai/yral-common.git", branch = "master" }
-yral-types = { git = "https://github.com/dolr-ai/yral-common.git", branch = "master" }
-yral-canisters-client = { git = "https://github.com/dolr-ai/yral-common.git", branch = "master", features = [
+], branch = "feat/update-post-details-utils-functions-to-fetch-username-from-metadata", version = "0.1.0" }
+sns-validation = { git = "https://github.com/dolr-ai/yral-common.git", branch = "feat/update-post-details-utils-functions-to-fetch-username-from-metadata" }
+yral-types = { git = "https://github.com/dolr-ai/yral-common.git", branch = "feat/update-post-details-utils-functions-to-fetch-username-from-metadata" }
+yral-canisters-client = { git = "https://github.com/dolr-ai/yral-common.git", branch = "feat/update-post-details-utils-functions-to-fetch-username-from-metadata", features = [
     "full",
 ] }
-yral-canisters-common = { git = "https://github.com/dolr-ai/yral-common.git", branch = "master" }
+yral-canisters-common = { git = "https://github.com/dolr-ai/yral-common.git", branch = "feat/update-post-details-utils-functions-to-fetch-username-from-metadata" }
 hon-worker-common = { git = "https://github.com/dolr-ai/yral-common.git", features = [
     "client",
-], branch = "master"}
-global-constants = { git = "https://github.com/dolr-ai/yral-common.git", branch = "master" }
-yral-identity = { git = "https://github.com/dolr-ai/yral-common.git", branch = "master", default-features = false, features = [
+], branch = "feat/update-post-details-utils-functions-to-fetch-username-from-metadata"}
+global-constants = { git = "https://github.com/dolr-ai/yral-common.git", branch = "feat/update-post-details-utils-functions-to-fetch-username-from-metadata" }
+yral-identity = { git = "https://github.com/dolr-ai/yral-common.git", branch = "feat/update-post-details-utils-functions-to-fetch-username-from-metadata", default-features = false, features = [
     "ic-git",
     "wasm-bindgen",
 ] }
-videogen-common = { git = "https://github.com/dolr-ai/yral-common.git", branch = "master", features = [
+videogen-common = { git = "https://github.com/dolr-ai/yral-common.git", branch = "feat/update-post-details-utils-functions-to-fetch-username-from-metadata", features = [
     "client",
     "ic",
 ] }
