+++ ssr/src/page/src/profile/mod.rs
@@ -68,7 +68,11 @@ struct TabsParam {
 }
 
 #[component]
-fn ListSwitcher1(user_canister: Principal, user_principal: Principal) -> impl IntoView {
+fn ListSwitcher1(
+    user_canister: Principal,
+    user_principal: Principal,
+    username: String,
+) -> impl IntoView {
     let param = use_params::<TabsParam>();
     let tab = Signal::derive(move || {
         param
@@ -125,7 +129,7 @@ fn ListSwitcher1(user_canister: Principal, user_principal: Principal) -> impl In
 
         <div class="flex flex-col gap-y-12 justify-center pb-12 w-11/12 sm:w-7/12">
             <Show when=move || current_tab() == 0>
-                <ProfilePosts user_canister user_principal />
+                <ProfilePosts user_canister user_principal username=username.clone()/>
             </Show>
             <Show when=move || current_tab() == 1>
                 <ProfileSpeculations user_canister user_principal />
@@ -217,7 +221,7 @@ fn ProfileViewInner(user: ProfileDetails) -> impl IntoView {
                         </Suspense>
                     </div>
                 </div>
-                <ListSwitcher1 user_canister user_principal />
+                <ListSwitcher1 user_canister user_principal username=username_or_fallback/>
             </div>
         </div>
     }
