+++ ssr/src/page/src/profile/posts.rs
@@ -97,7 +97,11 @@ fn Post(
 }
 
 #[component]
-pub fn ProfilePosts(user_principal: Principal, user_canister: Principal) -> impl IntoView {
+pub fn ProfilePosts(
+    user_principal: Principal,
+    user_canister: Principal,
+    username: String,
+) -> impl IntoView {
     let ProfilePostsContext {
         video_queue,
         start_index,
@@ -108,6 +112,7 @@ pub fn ProfilePosts(user_principal: Principal, user_canister: Principal) -> impl
         unauth_canisters(),
         video_queue,
         start_index,
+        username,
         user_principal,
         user_canister,
     );
